+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [#318](https://github.com/clojure-emacs/orchard/pull/318): **BREAKING:** Remove no longer used functions: `orchard.misc/lazy-seq?`, `orchard.misc/safe-count`, `orchard.misc/normalize-subclass`, `orchard.misc/remove-type-param`.
 * [#320](https://github.com/clojure-emacs/orchard/pull/320): Info: recognize printed Java classes/methods and munged Clojure functions in stacktrace outputs.
 * [#322](https://github.com/clojure-emacs/orchard/pull/322): Stacktrace: bring back `orchard.stacktrace` for stacktrace analysis (copied from `haystack.analyzer` and improved).
+* [#324](https://github.com/clojure-emacs/orchard/pull/324): Add dedicated renderers for exceptions to `orchard.print` and `orchard.inspect`.
 
 ## 0.30.1 (2025-02-24)